fixup local address detection
